Refactor dependency management and improve test setup
