+++ portals/developer-portal/it/rest-api/package.json
@@ -13,7 +13,9 @@
     "supertest": "^7.0.0",
     "cookiejar": "^2.1.4",
     "nock": "^13.5.5",
-    "better-sqlite3": "^11.3.0",
     "pg": "^8.14.0"
+  },
+  "optionalDependencies": {
+    "better-sqlite3": "^11.3.0"
   }
 }
