+++ portals/developer-portal/it/rest-api/subscriptions/webhook-events.spec.js
@@ -164,6 +164,8 @@ describe('subscriptions webhook events', () => {
     });
 
     it('encrypts secret fields to the subscriber public key when configured', async () => {
+        // TODO(pqc): migrate — RSA keypair mirrors the app's quantum-vulnerable
+        // envelope crypto (src/services/webhooks/envelopeCrypto.js); migrate together.
         const { publicKey, privateKey } = crypto.generateKeyPairSync('rsa', {
             modulusLength: 2048,
             publicKeyEncoding: { type: 'spki', format: 'pem' },
